2.WIP_Dataflow_PubSub/stream_container/src/requirements.txt
# Requirements
tweepy
google-cloud-pubsub
textblob